Moved package loader to common
